### 2024-07-05 #6824 @nstarman
@dfm, yes I'm very interested to review this JOSS submission. Could you confirm the timeline you need the review completed by? Looking forward to reading.
