### 2024-07-05 #6824 @dfm
Thanks @nstarman! Once the review begins, we typically try to complete the review in 4-6 weeks but, unlike your typical journal review, JOSS reviews are iterative without specific rounds of reviewing and responding, so it's useful to get started on the review ASAP. As soon as we have a second reviewer lined up I'll have a lot more specific info for everyone in a new thread, but I'm happy to answer any other questions in the meantime. Thanks again for agreeing to review!
