### 2024-07-16 #6824 @TomWagg
@dfm I'd be happy to review, looks very cool! (Though fair warning this would be my first time reviewing). However, I'm away on holiday until the start of August so I wouldn't be able to look properly until then.
