### 2024-07-19 ilikecubesnstuff/commensurability#16 @nstarman
📝: word choice
https://github.com/ilikecubesnstuff/commensurability/blob/ce0ab67a5ca17524eafd9c4fbfd5e9a4624a350e/paper/paper.md?plain=1#L35

I don't think the word "seems" is what you meant. How about just removing the word, to read
"One technique is “orbit tessellation”, which estimates the (sub-)volume of the toroid traversed by an orbit (Petersen et al., 2021)."

— openjournals/joss-reviews#7009
