### 2024-07-19 ilikecubesnstuff/commensurability#18 @nstarman
🔥 docs error
I'm trying to reproduce the example figure (which comes from https://commensurability.readthedocs.io/en/latest/tessellation/usage/analysis/rotating_bar/#adding-a-rotating-bar-to-a-potential) so I'm running that from the top.

I'm encountering the error:

```python
tanal = TessellationAnalysis2D(
    initial_condition,
    values,
    potential_definition,
    dt,
    steps,
    pattern_speed=omega,
    chunksize=500,
    mpchunksize=20,
)
```

produces the error `TypeError: MPAnalysisBase.__init__() got an unexpected keyword argument 'chunksize'`

Please update the docs to fix this example.

— openjournals/joss-reviews#7009
