### 2024-07-19 ilikecubesnstuff/commensurability#19 @nstarman
Test the docs as part of CI
Coming from #18, it would be great to test the docs as part of CI.
I recommend https://sybil.readthedocs.io/en/latest/, it's easy to use and very configurable.
I'm happy to push a PR.

— openjournals/joss-reviews#7009
