### 2024-07-19 ilikecubesnstuff/commensurability#20 @nstarman
Additional installation instructions 
`commensurability` requires `pidgey` (a dependency) which in turn requires a supported backend to be installed. However `pidgey` doesn't install any such backend (and shouldn't). I think there should be a small addition to the installation instructions discussing that while `commensurability` can work on pre-computed files and thus doesn't require any orbit integrator backend to function, in all likelihood a user will require such a library (agama, gala, galpy) to produce the file. Therefore I think there should be some mention of these backends and links to their installation pages.

— openjournals/joss-reviews#7009
