### 2024-08-21 #7009 @nstarman
Hi @ilikecubesnstuff! I would love to continue reviewing, but I'm finding https://github.com/ilikecubesnstuff/commensurability/issues/18 a blocker to reproducing the results. If you could post a fix in that Issue thread, even before merging a full PR, that would be appreciated. Thanks!
