### 2024-08-21 ilikecubesnstuff/commensurability#18 @ilikecubesnstuff
Changing `chunksize` and `mpchunksize` to `pidgey_chunksize` and `mp_chunksize` respectively should fix this error.
```py
tanal = TessellationAnalysis2D(
    initial_condition,
    values,
    potential_definition,
    dt,
    steps,
    pattern_speed=omega,
    pidgey_chunksize=500,
    mp_chunksize=20,
)
```

The chunk size parameters were renamed in 110e3c92ffeb08bac08299454bcf30bd48f719fa to make the package they are being passed to (`pidgey` vs `multiprocessing`) clearer, but it appears that the [corresponding pull request](https://www.github.com/ilikecubesnstuff/commensurability/pull/10) did not update the documentation to reflect these changes.

That being said, I am having issues installing Agama on my device (specifically, I updated my Agama version which broke the examples, but older versions throw errors during the installation process) so I cannot currently reproduce the figures either. I can look into updating the `readthedocs` code and figures with a more recent Agama version when I have time.
