+++ examples/using_agama/rotating_bar.py
@@ -49,12 +49,13 @@ def ic_function(x, vy, z):
     )
 
 
-dt = 0.01 * u.Gyr
-steps = 500
-tanal = TessellationAnalysis(
-    ic_function, values, pot, dt, steps, pattern_speed=omega, pidgey_chunksize=50
-)
-tanal.launch_interactive_plot("x", "vy")
+if __name__ == "__main__":
+    dt = 0.01 * u.Gyr
+    steps = 500
+    tanal = TessellationAnalysis(
+        ic_function, values, pot, dt, steps, pattern_speed=omega, pidgey_chunksize=50
+    )
+    tanal.launch_interactive_plot("x", "vy")
 
-tanal.save(f"examples/using_galpy/bar_{SIZE}_{FRAMES}.hdf5")
-tanal.launch_interactive_plot("x", "vy")
+    tanal.save(f"bar_{SIZE}_{FRAMES}.hdf5")
+    tanal.launch_interactive_plot("x", "vy")
