### 2024-08-26 ilikecubesnstuff/commensurability#24 @TomWagg
Quickstart error
Installation worked well but when I tried to run the quickstart code using `gala` I got the following error:
```
ValueError: Tessellation can only be performed on a 2D array of shape (npoints, ndim)
```

I'll copy the full traceback in a comment below. I wonder if the recent release of Gala v1.9 could have affected this?

— https://github.com/openjournals/joss-reviews/issues/7009
