### 2024-08-26 ilikecubesnstuff/commensurability#25 @TomWagg
Installation docs
I recommend adding some links to end of the [installation](https://commensurability.readthedocs.io/en/latest/installation/) docs where you mention agama/gala/galpy linking to their installation pages to save users needing to search for this.

Alternatively, you could even have another code cell at the bottom that demonstrates installing them and label it as optional.

— https://github.com/openjournals/joss-reviews/issues/7009
