### 2024-08-26 ilikecubesnstuff/commensurability#26 @TomWagg
Community guidelines
JOSS recommends that you includes community guidelines which help users to know how to:

- Contribute to the software
- Report issues or problems with the software
- Seek support

I think you could do this with another heading in your readme, or separate page in your docs pointing towards issues/PRs on GitHub. You might also consider creating [GitHub issue templates](https://docs.github.com/en/communities/using-templates-to-encourage-useful-issues-and-pull-requests/configuring-issue-templates-for-your-repository) to sort new issues easily?

— https://github.com/openjournals/joss-reviews/issues/7009
