### 2024-08-26 ilikecubesnstuff/commensurability#27 @TomWagg
Coverage of tests?
Have you tested the coverage of your test suite at all?

Tools like [coverage.py](https://coverage.readthedocs.io/en/7.6.1/index.html) can tell you what fraction of the code is actually covered by the tests to ensure you're not missing something. You can also set this up as part of the GitHub workflow when you do the tests.

— https://github.com/openjournals/joss-reviews/issues/7009
