### 2024-10-02 #7009 @ilikecubesnstuff
Hi all, I got extremely busy moving countries over September. I can continue working on this codebase now. Thank you for checking in!

The last time I opened this, my Agama installation was causing issues with my examples and automated tests. I plan to update the code to use Agama's latest version. However, in the interest of time, I can temporarily deprecate Agama support and remake the relevant documentation/examples with one of the other galactic dynamics packages if that is preferable.

Thank you for the additional issues @TomWagg, I will look through them shortly.
