### 2024-10-11 #7009 @ilikecubesnstuff
> The last time I opened this, my Agama installation was causing issues with my examples and automated tests. I plan to update the code to use Agama's latest version. However, in the interest of time, I can temporarily deprecate Agama support and remake the relevant documentation/examples with one of the other galactic dynamics packages if that is preferable.

I have isolated the version of Agama with which my examples were written (commit hash [0c5993d](https://github.com/GalacticDynamics-Oxford/Agama/tree/0c5993d1c631d9a9e8f48213f919e09bfd629639)). I can specify this commit hash in the README and relevant parts of the documentation - I'm not sure how else to work with Agama's versioning. 😅
