### 2024-10-11 4d56136
draft paper changes
