+++ paper/paper.md
@@ -28,35 +28,35 @@ arxiv-doi:
 
 # Summary
 
-Stars like our Sun orbit the center of our Milky Way galaxy. Over many orbits in the stellar disk of the galaxy, the Sun will (hypothetically) probe every point in a 3D toroid defined by a minimum/maximum radius and minimum/maximum height above the midplane. However, a special class of stars are on orbits that repeat their tracks with every revolution around the galactic center, and therefore only probe a small sub-volume of the toroid. These orbits are called “commensurate”. 
+Stars like our Sun orbit the center of our Milky Way galaxy. Over many orbits in the stellar disk of the galaxy, a star will (hypothetically) probe every point in a 3D toroid defined by a minimum/maximum radius and minimum/maximum height above the midplane. However, there are special classes of orbits that repeat their tracks with every revolution around the Galactic center, and therefore only probe a small sub-volume of the toroid. This property stems from low-integer orbital frequency ratios -- commensurate frequencies -- and such orbits are referred to as satisfying a "commensurability".
 
-To study the orbital content of galaxies, astronomers will often use models for galaxies that allow for the integration of orbits within the model potential. The integration results in a time series of positions (e.g. $[x,y,z]$) for the orbit that can then be analyzed to produce a classification. Complicating this picture, potentials evolving with time introduce a new frequency: the pattern frequency of the evolution. The evolution of the potential might be the rotation of a galactic bar, the formation or dissolution of spiral arms, the growth of the galaxy, or the interaction of a satellite galaxy. Orbits that are commensurate with the pattern frequency are particularly important in galactic dynamics as they experience the same force fluctuations during every revolution: this causes changes to the typically conserved quantities of orbits (e.g. energy and angular momentum).
+To study the orbital content of galaxies, astronomers often use models for galaxies that allow for the integration of orbits within the model galaxy's potential. The integration results in a time series of positions (e.g. $x,y,z$) for the orbit that can then be analyzed to produce a classification. Complicating this picture, potentials evolving with time introduce a new frequency: the pattern frequency of the evolution. The evolution of the potential might be sourced by the rotation of a galactic bar, the formation or dissolution of spiral arms, the growth of the galaxy, or the interaction with a satellite galaxy. Orbits that are commensurate with the pattern frequency are particularly important in galactic dynamics as they experience the same force fluctuations during every revolution: this causes changes to the typically conserved quantities of orbits (e.g. energy and angular momentum).
 
-Given their importance, astronomers have developed techniques to identify commensurate orbits. One technique is “orbit tessellation”, which seems to estimate the (sub-)volume of the toroid traversed by an orbit [@Petersen.Weinberg.Katz.2021]. Orbit tessellation aims to pick out commensurate orbits with relatively short orbit integration times, as well as pick out commensurabilities in cases where the kinematic frequencies need not stay constant (such as in a potential with multiple pattern frequencies from multiple causes).
+Given their importance, astronomers have developed techniques to identify commensurate orbits. One technique is “orbit tessellation”, which estimates the (sub-)volume of the toroid traversed by an orbit [@Petersen.Weinberg.Katz.2021]. The method of orbit tessellation aims to pick out commensurate orbits over relatively short orbit integration times, as well as pick out commensurabilities in cases where the kinematic frequencies need not stay constant (such as in a potential with multiple pattern frequencies from multiple causes).
 
-The Python package presented here, commensurability, provides a straightforward Python framework and connection to powerful tools for modeling potentials to estimate the volume of a toroid that a given orbit fills. 
+The Python package presented here, `commensurability`, provides a straightforward Python framework and connection to powerful tools for modeling potentials to estimate the volume of a toroid that a given orbit fills. 
[…]
