### 2024-10-11 ilikecubesnstuff/commensurability#16 @michael-petersen
Totally agree. This is addressed in #28.
