### 2024-10-11 ilikecubesnstuff/commensurability#17 @michael-petersen
Thanks again, these are great comments. I've made some proposed changes in #28; @ilikecubesnstuff can have a first pass, but @nstarman feel free to bounce back on anything in the PR!
