### 2024-10-12 fef96e4
additional draft paper edits
