+++ paper/paper.md
@@ -28,11 +28,11 @@ arxiv-doi:
 
 # Summary
 
-Stars like our Sun orbit the center of our Milky Way galaxy. Over many orbits in the stellar disk of the galaxy, a star will (hypothetically) probe every point in a 3D toroid defined by a minimum/maximum radius and minimum/maximum height above the midplane. However, there are special classes of orbits that repeat their tracks with every revolution around the Galactic center, and therefore only probe a small sub-volume of the toroid. This property stems from low-integer orbital frequency ratios -- commensurate frequencies -- and such orbits are referred to as satisfying a "commensurability".
+Stars like our Sun orbit the center of our Milky Way galaxy. Over many orbits in the stellar disk of the galaxy, a star will (hypothetically) probe every point in a 3D toroid defined by a minimum/maximum radius and minimum/maximum height above the midplane. However, there are special classes of orbits that repeat their tracks over many revolutions around the Galactic center, and therefore only probe a small sub-volume of the toroid. This property stems from low-integer orbital frequency ratios -- commensurate frequencies -- and such orbits are referred to as satisfying a "commensurability".
 
 To study the orbital content of galaxies, astronomers often use models for galaxies that allow for the integration of orbits within the model galaxy's potential. The integration results in a time series of positions (e.g. $x,y,z$) for the orbit that can then be analyzed to produce a classification. Complicating this picture, potentials evolving with time introduce a new frequency: the pattern frequency of the evolution. The evolution of the potential might be sourced by the rotation of a galactic bar, the formation or dissolution of spiral arms, the growth of the galaxy, or the interaction with a satellite galaxy. Orbits that are commensurate with the pattern frequency are particularly important in galactic dynamics as they experience the same force fluctuations during every revolution: this causes changes to the typically conserved quantities of orbits (e.g. energy and angular momentum).
 
-Given their importance, astronomers have developed techniques to identify commensurate orbits. One technique is “orbit tessellation”, which estimates the (sub-)volume of the toroid traversed by an orbit [@Petersen.Weinberg.Katz.2021]. The method of orbit tessellation aims to pick out commensurate orbits over relatively short orbit integration times, as well as pick out commensurabilities in cases where the kinematic frequencies need not stay constant (such as in a potential with multiple pattern frequencies from multiple causes).
+Given their importance, astronomers have developed techniques to identify commensurate orbits. One technique involves tessellating orbital coordinates to estimate the (sub-)volume of the toroid traversed [@Petersen.Weinberg.Katz.2021]. This "orbit tessellation" method aims to pick out commensurate orbits over relatively short orbit integration times, as well as pick out commensurabilities in cases where the kinematic frequencies need not stay constant (such as in a potential with multiple pattern frequencies from multiple causes).
 
 The Python package presented here, `commensurability`, provides a straightforward Python framework and connection to powerful tools for modeling potentials to estimate the volume of a toroid that a given orbit fills. 
 
@@ -52,7 +52,7 @@ Additionally, computing the orbit volume provides an opportunity to efficiently 
 
 `commensurability` provides several notable features, with an emphasis on extensibility and compatibility with existing code.
 
[…]
