### 2024-10-12 #7009 @ilikecubesnstuff
Addressing ilikecubesnstuff/commensurability#18, I have created a `joss` version of the docs hosted at https://commensurability.readthedocs.io/en/joss/.

This should contain any documentation changes in the `joss` branch (until it is merged into `main`).
