### 2024-10-14 ilikecubesnstuff/commensurability#24 @ilikecubesnstuff
Thank you for identifying this bug! This appears to be a problem within Pidgey, but it is also worth revising the default arguments for chunk sizes in commensurability.

When using chunk sizes greater than 1, Gala's integration routine takes in initial conditions `[ic1, ic2, ...]` and returns the integrated points as `[[p1_1, p2_1, ...], [p2_1, p2_2, ...], ...]`. However, supplying both `[ic]` and `ic` return `[p_1, p_2, ...]`. ilikecubesnstuff/pidgey#11 enforces the behavior of:
- `ic` --> `[p_1, p_2, ...]`
- `[ic]` --> `[[p_1], [p_2], ...]`
