### 2024-10-21 7440e73
:zap: Improve chunksize argument handling in analysis.py.
