+++ src/commensurability/analysis.py
@@ -32,6 +32,10 @@ from .evaluation import Evaluation
 from .interactive import InteractivePlot2D, InteractivePlot3D, InteractivePlotBase
 from .utils import collapse_coords, make_quantity
 
+# define default chunk size for orbit integration
+# unsure how necessary this is, revise exact value as needed
+DEFAULT_MAX_CHUNKSIZE = 1000
+
 
 class AnalysisBase(MappingABC):
     """
@@ -79,7 +83,7 @@ class AnalysisBase(MappingABC):
         pattern_speed: Union[float, u.Quantity] = 0.0,
         backend: Optional[Union[str, Backend]] = None,
         progressbar: bool = True,
-        pidgey_chunksize: int = 1,
+        pidgey_chunksize: int = None,
         _blank_measures: bool = False,
     ) -> None:
         """
@@ -130,8 +134,10 @@ class AnalysisBase(MappingABC):
         if not self.backend:
             raise TypeError(f"Unrecognized potential: {self.potential}")
 
+        if pidgey_chunksize is None:
+            pidgey_chunksize = min(int(self.size**0.5), DEFAULT_MAX_CHUNKSIZE)
         if pidgey_chunksize <= 0:
-            raise ValueError("chunksize must be greater than 0")
+            raise ValueError("pidgey_chunksize must be greater than 0")
         if pidgey_chunksize >= self.size:
             pidgey_chunksize = self.size
             # raise ValueError("chunksize must be less than total number of starting coordinates")
@@ -338,8 +344,8 @@ class MPAnalysisBase(AnalysisBase):
         pattern_speed: Union[float, u.Quantity] = 0.0,
         backend: Optional[Union[str, Backend]] = None,
         progressbar: bool = True,
-        pidgey_chunksize: int = 1,
-        mp_chunksize: int = 1,
+        pidgey_chunksize: int = None,
+        mp_chunksize: int = None,
         _blank_measures: bool = False,
     ) -> None:
         super().__init__(
@@ -354,6 +360,23 @@ class MPAnalysisBase(AnalysisBase):
             pidgey_chunksize=pidgey_chunksize,
             _blank_measures=True,
         )
+        if pidgey_chunksize is None:
+            # set default chunk size for pidgey if not provided
+            pidgey_chunksize = min(int(self.size**0.5), DEFAULT_MAX_CHUNKSIZE)
+        if pidgey_chunksize <= 0:
+            raise ValueError("pidgey_chunksize must be greater than 0")
+        if pidgey_chunksize >= self.size:
+            pidgey_chunksize = self.size
+            # raise ValueError("chunksize must be less than total number of starting coordinates")
+
+        if mp_chunksize is None:
+            # set default chunk size for multiprocessing if not provided
+            mp_chunksize = int(pidgey_chunksize**0.5)
+        if mp_chunksize <= 0:
+            raise ValueError("mp_chunksize must be greater than 0")
+        if mp_chunksize > pidgey_chunksize:
+            raise ValueError("mp_chunksize must not be greater than pidgey_chunksize")
+
         if not _blank_measures:
             self._construct_image_with_mp(pidgey_chunksize, mp_chunksize, progressbar)
 
@@ -384,6 +407,7 @@ class MPAnalysisBase(AnalysisBase):
                 values = list(
                     tqdm(
                         p.imap(self.__eval__, orbits, chunksize=mp_chunksize),
+                        desc=f"with {mp_chunksize=}",
                         total=pidgey_chunksize,
                         leave=False,
                     )
