### 2024-10-21 2cdaaa6
:rotating_light: Fix type-hinting issues.
