### 2024-11-03 2217abb
:construction_worker: Add coverage reporting. Update linting and type-hinting tox environments to only run once. Add coverage reporting with pytest-cov to tox scripts. Add coverage reporting with Coveralls in GitHub Actions. Update lock file.
