+++ pyproject.toml
@@ -56,6 +56,7 @@ dev = [
 ]
 test = [
     "pytest>=8.2.0",
+    "pytest-cov>=6.0.0",
 ]
 type = [
     "mypy>=1.10.0",
