+++ .github/workflows/tests.yml
@@ -7,7 +7,7 @@ on:
   workflow_dispatch:
 
 jobs:
-  tox:
+  test:
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
@@ -25,3 +25,19 @@ jobs:
     - name: Test with tox
       run: |
         pdm run -v tox
+    - name: Coveralls Parallel
+      uses: coverallsapp/github-action@v2
+      with:
+        flag-name: run-${{ matrix.os }}-${{ matrix.python-version }}
+        parallel: true
+
+coverage:
+  needs: test
+  if: ${{ always() }}
+  runs-on: ubuntu-latest
+  steps:
+  - name: Coveralls
+    uses: coverallsapp/github-action@v2
+    with:
+      parallel-finished: true
+      carryforward: "run-windows-latest-3.11,run-windows-latest-3.12,run-ubuntu-latest-3.11,run-ubuntu-latest-3.12,run-macos-latest-3.11,run-macos-latest-3.12"
