+++ tox.ini
@@ -1,15 +1,21 @@
 [tox]
-env_list = py{311, 312}, type, lint
+env_list = py{311, 312}-{linux,macos,windows}, type, lint
 
 [gh-actions]
 python =
     3.11: py311, type, lint
     3.12: py312
 
+[gh-actions:env]
+PLATFORM =
+    ubuntu-latest: linux, type, lint
+    macos-latest: macos
+    windows-latest: windows
+
 [testenv]
 groups = test
 commands =
-    pytest tests
+    pytest tests --cov --cov-append
 
 [testenv:type]
 groups = type
