+++ .github/workflows/tests.yml
@@ -31,13 +31,13 @@ jobs:
         flag-name: run-${{ matrix.os }}-${{ matrix.python-version }}
         parallel: true
 
-coverage:
-  needs: test
-  if: ${{ always() }}
-  runs-on: ubuntu-latest
-  steps:
-  - name: Coveralls
-    uses: coverallsapp/github-action@v2
-    with:
-      parallel-finished: true
-      carryforward: "run-windows-latest-3.11,run-windows-latest-3.12,run-ubuntu-latest-3.11,run-ubuntu-latest-3.12,run-macos-latest-3.11,run-macos-latest-3.12"
+  coverage:
+    needs: test
+    if: ${{ always() }}
+    runs-on: ubuntu-latest
+    steps:
+    - name: Coveralls
+      uses: coverallsapp/github-action@v2
+      with:
+        parallel-finished: true
+        carryforward: "run-windows-latest-3.11,run-windows-latest-3.12,run-ubuntu-latest-3.11,run-ubuntu-latest-3.12,run-macos-latest-3.11,run-macos-latest-3.12"
