### 2024-11-04 ecaa9c3
:green_heart: Adapt tox script from [pytest-cov](https://pytest-cov.readthedocs.io/en/latest/reporting.html).
