+++ tox.ini
@@ -15,7 +15,7 @@ PLATFORM =
 [testenv]
 groups = test
 commands =
-    pytest tests --cov --cov-append
+    pytest --cov-report= --cov=src tests
 
 [testenv:type]
 groups = type
