### 2024-11-04 4563f95
:green_heart: Add coverage to the dev dependency group.
