+++ pyproject.toml
@@ -33,26 +33,11 @@ requires = [
 ]
 build-backend = "pdm.backend"
 
-[tool.black]
-line-length = 100
-target-version = [
-    "py311",
-    "py312",
-]
-
-[tool.isort]
-profile = "black"
-
-[tool.mypy]
-ignore_missing_imports = true
-
-[tool.pdm]
-distribution = true
-
-[tool.pdm.dev-dependencies]
+[dependency-groups]
 dev = [
     "tox-pdm>=0.7.2",
     "tox>=4.15.0",
+    "coverage>=7.6.4"
 ]
 test = [
     "pytest>=8.2.0",
@@ -76,3 +61,19 @@ docs = [
 github = [
     "tox-gh-actions>=3.2.0",
 ]
+
+[tool.black]
+line-length = 100
+target-version = [
+    "py311",
+    "py312",
+]
+
+[tool.isort]
+profile = "black"
+
+[tool.mypy]
+ignore_missing_imports = true
+
+[tool.pdm]
+distribution = true
