+++ tox.ini
@@ -1,16 +1,10 @@
 [tox]
-env_list = py{311, 312}-{linux,macos,windows}, type, lint
+env_list = type, lint, py{311, 312}
 
 [gh-actions]
 python =
     3.11: py311, type, lint
     3.12: py312
-
-[gh-actions:env]
-PLATFORM =
-    ubuntu-latest: linux, type, lint
-    macos-latest: macos
-    windows-latest: windows
 
 [testenv]
 groups = test
