+++ tox.ini
@@ -7,6 +7,7 @@ python =
     3.12: py312
 
 [testenv]
+usedevelop = True
 groups = test
 commands =
     pytest --cov-report= --cov=src tests
