+++ .github/workflows/tests.yml
@@ -25,15 +25,11 @@ jobs:
     - name: Test with tox
       run: |
         pdm run -v tox
-    - name: Prepare Coverage File
-      run: |
-        pdm run -v coverage xml
-        cat coverage.xml
     - name: Coveralls Parallel
       uses: coverallsapp/github-action@v2
       with:
         flag-name: run-${{ matrix.os }}-${{ matrix.python-version }}
-        format: cobertura
+        format: python
         parallel: true
 
   coverage:
