### 2024-11-04 a957afe
:hammer: Add coverage configuration to pyproject.toml.
