### 2024-11-04 fae3b15
:white_check_mark: Fix main tests for analysis class.
