+++ tests/test_analysis.py
@@ -1,5 +1,6 @@
-import numpy as np
 import pytest
+from astropy import coordinates as c
+from astropy import units as u
 from pidgey.base import Backend
 
 from commensurability import TessellationAnalysis, TessellationAnalysis2D
@@ -12,17 +13,28 @@ def dummy_backend():
             return list
 
         def _compute_orbit(self, skycoord, pot, dt, steps, pattern_speed):
-            return [(0, 0, 0), (0, 0, 1), (0, 1, 0), (1, 0, 0)]
+            return c.SkyCoord(
+                x=[[0], [0], [0], [1]],
+                y=[[0], [0], [1], [0]],
+                z=[[0], [1], [0], [0]],
+                unit='kpc',
+                representation_type='cartesian'
+            )
 
         def _extract_points(self, orbit, pattern_speed):
-            return [(0, 0, 0), (0, 0, 1), (0, 1, 0), (1, 0, 0)]
+            return orbit.data.T
 
     return DummyBackend()
 
 
 @pytest.fixture
 def ic_params():
-    return lambda x, y: (x, y), {"x": [0, 1, 2], "y": [0, 1, 2]}
+    ic_function = lambda x, y: c.SkyCoord(
+        x=x * u.kpc, y=y * u.kpc, z=0 * u.kpc,
+        v_x=0 * u.km/u.s, v_y=0 * u.km/u.s, v_z=0 * u.km/u.s,
+        representation_type='cartesian'
+    )
+    return ic_function, {"x": [0, 1, 2], "y": [0, 1, 2]}
 
 
 @pytest.fixture
@@ -90,7 +102,6 @@ class TestAnalysis:
             1,
             1,
             backend=dummy_backend,
-            _blank_measures=True,
         )
         assert analysis.ic_function == ic_function
         assert analysis.ic_values == ic_values
@@ -103,7 +114,6 @@ class TestAnalysis:
         for pixel in analysis:
             ic, measure = analysis[pixel]
             # specifically for these dummy examples
-            assert ic == pixel
             assert measure == 0.0
 
 
@@ -169,7 +179,6 @@ class TestAnalysis2D:
             1,
             1,
             backend=dummy_backend,
-            _blank_measures=True,
         )
         assert analysis.ic_function == ic_function
         assert analysis.ic_values == ic_values
@@ -182,5 +191,4 @@ class TestAnalysis2D:
         for pixel in analysis_2d:
             ic, measure = analysis_2d[pixel]
             # specifically for these dummy examples
-            assert ic == pixel
             assert measure == 0.0
