### 2024-11-04 10f8177
:memo: Add coverage badges to home pages. :memo: Update license badges in home pages.
