+++ README.md
@@ -2,10 +2,12 @@
 
 [![PyPI - Version](https://img.shields.io/pypi/v/commensurability)](https://pypi.org/project/commensurability/)
 [![PyPI - Python Version](https://img.shields.io/pypi/pyversions/commensurability)](https://pypi.org/project/commensurability/)
+[![PyPI - License](https://img.shields.io/pypi/l/commensurability)](https://github.com/ilikecubesnstuff/commensurability/blob/main/LICENSE)
 [![pdm-managed](https://img.shields.io/badge/pdm-managed-blueviolet)](https://pdm.fming.dev)
 [![pre-commit](https://img.shields.io/badge/pre--commit-enabled-brightgreen?logo=pre-commit)](https://github.com/pre-commit/pre-commit)
 [![Code style: black](https://img.shields.io/badge/code%20style-black-000000.svg)](https://github.com/psf/black)
 [![Imports: isort](https://img.shields.io/badge/%20imports-isort-%231674b1?style=flat&labelColor=ef8336)](https://pycqa.github.io/isort/)
+[![Coverage Status](https://coveralls.io/repos/github/ilikecubesnstuff/commensurability/badge.svg?branch=main)](https://coveralls.io/github/ilikecubesnstuff/commensurability?branch=main)
 
 A Python package for performing analysis on orbit commensurabilities.
 
