### 2024-11-04 681034b
:green_heart: Rename custom PDM command to 'covreport'.
