### 2024-11-04 ilikecubesnstuff/commensurability#27 @ilikecubesnstuff
Thank you for the suggestion! The code coverage for this package should now be available on [Coveralls](https://coveralls.io/github/ilikecubesnstuff/commensurability) here: https://coveralls.io/github/ilikecubesnstuff/commensurability

There are a few parts of the code I am unsure how to cover with tests, and have chosen to ignore in the coverage reports:
1. matplotlib code (particularly the interactive plots with keyboard/mouse inputs).
2. code specific to supported backends (like `agama`/`gala`/`galpy`).

Otherwise, the current code coverage is ~71%. I will merge #30 after writing a few more tests to improve the coverage.
