### 2024-11-04 ilikecubesnstuff/commensurability#19 @ilikecubesnstuff
Thank you for the suggestion, this is certainly a good idea. There are a few concerns I have regarding using [Sybil](https://sybil.readthedocs.io/en/latest/index.html):
1. Is the markdown parser configurable for files such as [`mw_orbit.md`](https://github.com/ilikecubesnstuff/commensurability/blob/joss/docs/tessellation/usage/tessellation/mw_orbit.md?plain=1)? In the [documentation](https://commensurability.readthedocs.io/en/latest/tessellation/usage/tessellation/mw_orbit/), this formatting displays multiple tabs above each code block.
2. Many of the documentation scripts depend on packages like `agama` or `galpy`, sometimes specific versions; I am unsure how to extract scripts particular to a certain package with [Sybil](https://sybil.readthedocs.io/en/latest/index.html). Ideally, all the scripts associated with one package happen in one testing environment, but configuring this into CI is unclear to me.

Pushing a PR would be greatly appreciated, please let me know if that can be arranged!
