### 2024-11-05 c526ea5
:memo: Add basic contributing guidelines.
