+++ README.md
@@ -25,3 +25,11 @@ See the [documentation](https://commensurability.readthedocs.io/en/latest/) page
 
 This package is to be used along with one of the galactic dynamics packages [`agama`](https://github.com/GalacticDynamics-Oxford/Agama), [`gala`](https://gala-astro.readthedocs.io/en/latest/), or [`galpy`](https://docs.galpy.org/en/latest/).
 There are scripts for each under the `examples` folder that can be run to test whether everything is working correctly.
+
+## Contributing
+
+If you wish to contribute to ``commensurability``, please fork the repository and create a [pull request](https://github.com/ilikecubesnstuff/commensurability/pulls).
+
+If you wish to report bugs, request features or suggest other ideas, please open an [issue](https://github.com/ilikecubesnstuff/commensurability/issues).
+
+For more information, see [contributing] in the documentation.
