### 2024-11-06 ilikecubesnstuff/commensurability#27 @TomWagg
Great! Yeah testing plotting and dependency code can be tricky. In some cases it may be appropriate to exclude some of the code from the coverage test: https://coverage.readthedocs.io/en/7.6.4/excluding.html
