### 2024-11-06 bea736f
:bug: Fix indentation bug in serialization. :bug: Fix generic simplex sides bug. :memo: Add docstring for missing method sentinel.
