### 2024-11-06 ef097cf
:wrench: Improve excluding plotting code from coverage.
