### 2024-11-06 f05c056
:rotating_light: Fix code formatting with black.
