### 2024-11-11 aa7a05a
ci: ensure astropy installed in test env
Signed-off-by: nstarman <nstarman@users.noreply.github.com>
