+++ pyproject.toml
@@ -42,6 +42,7 @@ dev = [
 test = [
     "pytest>=8.2.0",
     "pytest-cov>=6.0.0",
+    "astropy>=6.0.0"
 ]
 type = [
     "mypy>=1.10.0",
