### 2024-11-11 d9b65f1
ci: run pre-commit
Signed-off-by: nstarman <nstarman@users.noreply.github.com>
