+++ paper/paper.bib
@@ -88,15 +88,15 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@article{qhull.paper, 
-        author = {Barber, C. Bradford and Dobkin, David P. and Huhdanpaa, Hannu}, 
-        title = {The quickhull algorithm for convex hulls}, 
-        year = {1996}, 
-        issue_date = {Dec. 1996}, 
-        publisher = {Association for Computing Machinery}, 
-        address = {New York, NY, USA}, volume = {22}, number = {4}, issn = {0098-3500}, url = {https://doi.org/10.1145/235815.235821}, doi = {10.1145/235815.235821}, 
-        journal = {ACM Trans. Math. Softw.}, month = {dec}, pages = {469–483}, numpages = {15}, 
-        keywords = {halfspace intersection, convex hull, Voronoi diagram, Delaunay triangulation} 
+@article{qhull.paper,
+        author = {Barber, C. Bradford and Dobkin, David P. and Huhdanpaa, Hannu},
+        title = {The quickhull algorithm for convex hulls},
+        year = {1996},
+        issue_date = {Dec. 1996},
+        publisher = {Association for Computing Machinery},
+        address = {New York, NY, USA}, volume = {22}, number = {4}, issn = {0098-3500}, url = {https://doi.org/10.1145/235815.235821}, doi = {10.1145/235815.235821},
+        journal = {ACM Trans. Math. Softw.}, month = {dec}, pages = {469–483}, numpages = {15},
+        keywords = {halfspace intersection, convex hull, Voronoi diagram, Delaunay triangulation}
 }
 
 @ARTICLE{scipy.paper,
