+++ tests/tessellation/test_constructor.py
@@ -7,21 +7,22 @@ import pytest
 
 from commensurability.tessellation import Tessellation
 
+
 class TestTessellationConstructor:
-    
     def test_list_subclass(self):
-        
         class other_list(list):
             pass
-        
-        points = other_list([
-            [1, 0, 0],
-            [-1, 0, 0],
-            [0, 1, 0],
-            [0, -1, 0],
-            [0, 0, 1],
-            [0, 0, -1],
-        ])
+
+        points = other_list(
+            [
+                [1, 0, 0],
+                [-1, 0, 0],
+                [0, 1, 0],
+                [0, -1, 0],
+                [0, 0, 1],
+                [0, 0, -1],
+            ]
+        )
         tess = Tessellation(points, axis_ratio=np.inf, qhull_options="Qz", incremental=False)
         assert math.isclose(tess.measure, 1)
 
@@ -32,9 +33,9 @@ class TestTessellationConstructor:
 
     def test_cartersian_representation(self):
         points = c.CartesianRepresentation(
-            x = [1, -1, 0, 0, 0, 0] * u.kpc,
-            y = [0, 0, 1, -1, 0, 0] * u.kpc,
-            z = [0, 0, 0, 0, 1, -1] * u.kpc,
+            x=[1, -1, 0, 0, 0, 0] * u.kpc,
+            y=[0, 0, 1, -1, 0, 0] * u.kpc,
+            z=[0, 0, 0, 0, 1, -1] * u.kpc,
         )
         tess = Tessellation(points, axis_ratio=np.inf, qhull_options="Qz", incremental=False)
         assert math.isclose(tess.measure, 1)
