+++ tests/tessellation/test_generic.py
@@ -5,15 +5,10 @@ import pytest
 
 from commensurability.tessellation import Tessellation
 
-class TestNormalizations:
 
+class TestNormalizations:
     @pytest.mark.parametrize(
-        'norm,ev', 
-        [
-            ('nsphere_approx', 0),
-            ('convexhull', 1),
-            ('convexhull_rot4', 0)
-        ]
+        "norm,ev", [("nsphere_approx", 0), ("convexhull", 1), ("convexhull_rot4", 0)]
     )
     def test_sphere(self, norm, ev):
         points = [
@@ -23,7 +18,13 @@ class TestNormalizations:
             [1000, 0, 1, 0],
             [1000, 0, 0, 1],
         ]
-        tess = Tessellation(points, normalization_routine=norm, axis_ratio=np.inf, qhull_options="Qz", incremental=False)
+        tess = Tessellation(
+            points,
+            normalization_routine=norm,
+            axis_ratio=np.inf,
+            qhull_options="Qz",
+            incremental=False,
+        )
         assert math.isclose(tess.measure, ev, abs_tol=1e-6)
 
     def test_unrecognized_normalization(self):
@@ -35,4 +36,10 @@ class TestNormalizations:
             [1000, 0, 0, 1],
         ]
         with pytest.raises(ValueError, match="Unrecognized normalization routine"):
-            Tessellation(points, normalization_routine="unknown", axis_ratio=np.inf, qhull_options="Qz", incremental=False)
+            Tessellation(
+                points,
+                normalization_routine="unknown",
+                axis_ratio=np.inf,
+                qhull_options="Qz",
+                incremental=False,
+            )
