### 2024-11-11 d39ec2c
feat: doctests
Signed-off-by: nstarman <nstarman@users.noreply.github.com>
