### 2024-11-11 944190b
ci: add sybil to test
Signed-off-by: nstarman <nstarman@users.noreply.github.com>
