### 2024-11-11 4b6f316
ci: try putting sybil in dev.
Signed-off-by: nstarman <nstarman@users.noreply.github.com>
