+++ pyproject.toml
@@ -37,7 +37,8 @@ build-backend = "pdm.backend"
 dev = [
     "tox-pdm>=0.7.2",
     "tox>=4.15.0",
-    "coverage>=7.6.4"
+    "coverage>=7.6.4",
+    "sybil>=8.0.1",
 ]
 test = [
     "pytest>=8.2.0",
