### 2024-11-30 ce48930
build: add astropy dependency
Signed-off-by: nstarman <nstarman@users.noreply.github.com>
