### 2024-11-30 8ac7f1b
fix: add required equivalency
This fix should be upstreamed to pidgey and then removed here.

Signed-off-by: nstarman <nstarman@users.noreply.github.com>
