### 2024-11-30 ca84c20
refactor: pre-allocate coords list size
Signed-off-by: nstarman <nstarman@users.noreply.github.com>
