### 2024-11-30 ilikecubesnstuff/commensurability#35 @nstarman
Galpy orbit initialization
I'm seeing:
`galpyWarning: Orbit's initialization normalization ro and zo are incompatible with SkyCoord's galcen_distance (should have galcen_distance^2 = ro^2 + zo^2)`

ro and zo can be passed to galpy's constructors / methods, and I guess pidgey should manage that. It can cause issues when doing frame transforms from galpy.

— https://github.com/openjournals/joss-reviews/issues/7009
