### 2024-11-30 ilikecubesnstuff/commensurability#36 @nstarman
Missing required `ax` arg
These are missing required args.

https://github.com/ilikecubesnstuff/commensurability/blob/7c25c9e34690169c76f3f6387c565440f3144279/docs/tessellation/usage/tessellation/pidgey.md?plain=1#L79

https://github.com/ilikecubesnstuff/commensurability/blob/7c25c9e34690169c76f3f6387c565440f3144279/docs/tessellation/usage/tessellation/mw_orbit.md?plain=1#L101

— https://github.com/openjournals/joss-reviews/issues/7009
