### 2024-11-30 ilikecubesnstuff/commensurability#37 @nstarman
Incorrect imports
There are still a number of `from tessellation import Tessellation`

E.g.

https://github.com/ilikecubesnstuff/commensurability/blob/7c25c9e34690169c76f3f6387c565440f3144279/docs/tessellation/usage/tessellation/mw_orbit.md?plain=1#L4

— https://github.com/openjournals/joss-reviews/issues/7009
