### 2024-11-30 ilikecubesnstuff/commensurability#38 @nstarman
Missing astropy.units equivalency in the docs
The astropy equivalency `dimensionless_angles` isn't always set. This is required by

https://github.com/ilikecubesnstuff/commensurability/blob/7c25c9e34690169c76f3f6387c565440f3144279/docs/tessellation/usage/tessellation/pidgey.md?plain=1#L65

This should be fixed in pidgey and downstreamed to here.

— https://github.com/openjournals/joss-reviews/issues/7009
