### 2024-12-01 ilikecubesnstuff/commensurability#39 @nstarman
incorrect kwargs in docs
https://github.com/ilikecubesnstuff/commensurability/blob/7c25c9e34690169c76f3f6387c565440f3144279/docs/tessellation/usage/analysis/rotating_bar.md?plain=1#L61

https://github.com/ilikecubesnstuff/commensurability/blob/7c25c9e34690169c76f3f6387c565440f3144279/docs/tessellation/usage/analysis/solar_neighborhood.md?plain=1#L55

— https://github.com/openjournals/joss-reviews/issues/7009
