### 2024-12-01 #7009 @nstarman
Thanks @ilikecubesnstuff for updating the docs!
I ran all the galpy portions of the docs and found only a few small issues (see above Issues / PRs).
IMO the only blocker to this paper is full doc tests, which would ensure that the docs examples I ran would work without issue. I made a draft PR https://github.com/ilikecubesnstuff/commensurability/pull/32 to help get this set up. I really like how the docs demonstrate how to use each of `galpy`, `agama`, `gala`, but this definitely does make doing doc tests more challenging. In https://github.com/ilikecubesnstuff/commensurability/pull/32 I discuss how this can be done. IMO doc tests are the most important tests as users will copy-paste examples directly from the docs. If the examples don't work...
