### 2024-12-15 #7009 @ilikecubesnstuff
Thank you so much for the added comments! The open PRs and PR draft are massively helpful.

I think that the documentation requires a little restructuring to be docile to existing doctest tools. I will continue the conversation in https://github.com/ilikecubesnstuff/commensurability/pull/32 as I work on it.

The coming weeks are quite busy for me, but I will try to make some progress over my winter break.
