### 2025-01-29 ed3f74a
:pushpin: Update lock file & remove redundant dependencies.
