+++ docs/tessellation/usage/tessellation/multiprocessing.md
@@ -1,3 +1,32 @@
+<!-- invisible-code-block: python
+try:
+    import agama
+    AGAMA_AVAILABLE = True
+except ModuleNotFoundError:
+    AGAMA_AVAILABLE = False
+
+try:
+    import gala
+    GALA_AVAILABLE = True
+except ModuleNotFoundError:
+    GALA_AVAILABLE = False
+
+try:
+    import galpy
+    GALPY_AVAILABLE = True
+except ModuleNotFoundError:
+    GALPY_AVAILABLE = False
+
+# NOTE: agama potential is not defined in scope, skip even if agama is available
+AGAMA_AVAILABLE = False
+
+# guarantee that exactly one is used
+USE_AGAMA = AGAMA_AVAILABLE
+USE_GALA = not AGAMA_AVAILABLE and GALA_AVAILABLE
+USE_GALPY = not AGAMA_AVAILABLE and not GALA_AVAILABLE and GALPY_AVAILABLE
+RUN = USE_AGAMA or USE_GALA or USE_GALPY
+-->
+
 # Using Multiprocessing
 
 This tutoral is similar to [Milky Way Orbit](mw_orbit.md) and [Using Pidgey](pidgey.md).
@@ -7,6 +36,8 @@ However, this will explore executing these routines on multiple orbits at once u
 ```python
 from multiprocessing import Pool
 
+import numpy as np
+
 from commensurability.tessellation import Tessellation
 ```
 
@@ -14,6 +45,8 @@ Define a Milky Way potential with your package of choice.
 Initialize the corresponding backend with `pidgey`.
 
 === "Agama"
+
+    <!-- skip: next if(not USE_AGAMA) -->
 
     ```python
     import pidgey
@@ -31,6 +64,8 @@ Initialize the corresponding backend with `pidgey`.
 
 === "Gala"
 
+    <!-- skip: next if(not USE_GALA) -->
+
     ```python
     import pidgey
     backend = pidgey.GalaBackend()
@@ -40,6 +75,8 @@ Initialize the corresponding backend with `pidgey`.
     ```
 
 === "Galpy"
+
+    <!-- skip: next if(not USE_GALPY) -->
 
     ```python
     import pidgey
@@ -51,6 +88,8 @@ Initialize the corresponding backend with `pidgey`.
 
 Define initial conditions using [`astropy`](https://www.astropy.org/) and perform the orbit integration routine.
 Here a list of $N=500$ orbits is defined.
+
+<!-- skip: next if(not RUN) -->
 
 ```python
 import astropy.coordinates as c
@@ -75,12 +114,16 @@ orbits = backend.compute_orbit(ics, potential, 0.01 * u.Gyr, 200)
 To speed up the commensurability evaluation, we can use [`multiprocessing.Pool`](https://docs.python.org/3/library/multiprocessing.html#multiprocessing.pool.Pool) (specifically [`multiprocessing.Pool.map`](https://docs.python.org/3/library/multiprocessing.html#multiprocessing.pool.Pool.map)) to parallelize tessellation.
 First, we have to define a function to map over all orbit points.
 
+<!-- skip: next if(not RUN) -->
+
 ```python
 def evaluate(points):
     return Tessellation(points).measure
 ```
 
 Then within a `Pool()` context, map `Tessellation` over the orbit points.
+
+<!-- skip: next "multiprocessing does not work here, evaulate is not pickleable" -->
 
 ```python
 point_sets = orbits.xyz.transpose(1, 2, 0)  # change to shape (norbits, npoints, ndims)
@@ -90,6 +133,8 @@ with Pool() as p:
 
 After leaving this running, a list of commensurability values should remain.
[…]
