### 2025-02-15 5c66d2b
:arrow_up: Update depedency groups and lock file. Move astropy into required dependencies. Remove "doctests" group, combine into tests - sybil is required for tests regardless.
