### 2025-02-15 40846bf
:construction_worker: Update tests workflow. Use miniconda for setup, to match the cron-based workflows.
