+++ .github/workflows/tests.yml
@@ -12,7 +12,7 @@ jobs:
     strategy:
       matrix:
         os: [ubuntu-latest, windows-latest, macos-latest]
-        python-version: ["3.11", "3.12"]
+        python-version: [3.11, 3.12]
     steps:
     - uses: actions/checkout@v4
     - name: Setup PDM
@@ -50,12 +50,32 @@ jobs:
     if: github.event_name == 'workflow_dispatch' || contains(github.event.pull_request.labels.*.name, 'doctests')
     runs-on: ubuntu-latest
     steps:
-    - uses: actions/checkout@v4
-    - name: Setup PDM
-      uses: pdm-project/setup-pdm@v4
-    - name: Install dependencies
+    - name: Checkout
+      uses: actions/checkout@v4
+
+    - name: Setup Miniconda with Python ${{ matrix.python-version }}
+      uses: conda-incubator/setup-miniconda@v3
+      with:
+        auto-update-conda: true
+        python-version: ${{ matrix.python-version }}
+
+    - name: Install Packages
+      # "bash -el {0}" required with conda-incubator/setup-miniconda@v3
+      # see: https://github.com/conda-incubator/setup-miniconda?tab=readme-ov-file#important
+      shell: bash -el {0}
+      # test environment is set up by conda-incubator/setup-miniconda@v3
+      # dependencies to install:
+      # - galpy: the orbit integration backend
+      # - gsl: used by galpy for fast integration
+      # - pytest: testing framework to run the test scripts
       run: |
-        pdm sync -G dev, -G github, -G doctest
-    - name: Run doctests
-      run: |
-        pdm run -v pytest docs
+        conda activate test
+        conda install -c conda-forge gsl galpy pytest sybil
+        python -m pip install --upgrade pip
+        python -m pip install .
+
+    - name: Test Docs
+      # "bash -el {0}" required with conda-incubator/setup-miniconda@v3
+      # see: https://github.com/conda-incubator/setup-miniconda?tab=readme-ov-file#important
+      shell: bash -el {0}
+      run: pytest docs
