### 2025-02-15 560353a
:arrow_down: Downgrade astropy dependency.
