+++ pyproject.toml
@@ -12,7 +12,7 @@ dependencies = [
     "h5py>=3.10.0",
     "more_itertools>=10.2.0",
     "scipy>=1.13.0",
-    "astropy>=7.0.1",
+    "astropy>=6.1.0",
 ]
 classifiers = [
     "Intended Audience :: Science/Research",
