### 2025-02-15 713afc0
:construction_worker: Run doctests on Python 3.12.
