### 2025-02-16 #7009 @ilikecubesnstuff
Hi @nstarman, are https://github.com/ilikecubesnstuff/commensurability/pull/33 and https://github.com/ilikecubesnstuff/commensurability/pull/34 meant to be merged into `main`?

I realized now that both changes have already been addressed in the `joss` branch. Would it make more sense to close the pull requests and link to the corresponding changes, or is there a reason to merge these directly into `main`?
