### 2025-02-17 #7009 @nstarman
> Hi @nstarman, are https://github.com/ilikecubesnstuff/commensurability/pull/33 and https://github.com/ilikecubesnstuff/commensurability/pull/34 meant to be merged into main?

No reason so long as the `joss` branch is merged into `main`, which I'm sure is the intent 👍.
