### 2025-02-19 8bb3c6c
:arrow_up: Update pidgey to 1.0.3 and revert commensurability changes.
