+++ src/commensurability/analysis.py
@@ -366,21 +366,20 @@ class MPAnalysisBase(AnalysisBase):
             total=self.size // pidgey_chunksize,
             disable=not progressbar,
         ):
-            coords = [
-                self.ic_function(*(self.ic_values[ax][i] for i, ax in zip(pixel, self.axis_names)))
-                for pixel in pixels
-            ]
+            coords = []
+            for pixel in pixels:
+                params = [self.ic_values[ax][i] for i, ax in zip(pixel, self.axis_names)]
+                coord = self.ic_function(*params)
+                coords.append(coord)
             coords = collapse_coords(coords)
 
-            # TODO: this is a hack that should be fixed in pidgey
-            with u.add_enabled_equivalencies(u.dimensionless_angles()):
-                orbits = self.backend.compute_orbit(
-                    coords,
-                    self.potential,
-                    self.dt,
-                    self.steps,
-                    pattern_speed=self.pattern_speed,
-                )
+            orbits = self.backend.compute_orbit(
+                coords,
+                self.potential,
+                self.dt,
+                self.steps,
+                pattern_speed=self.pattern_speed,
+            )
             with Pool() as p:
                 values = list(
                     tqdm(
