### 2025-02-20 586c578
:arrow_up: Upgrade pidgey to 1.0.3. Reflecting ilikecubesnstuff/commensurability#34.
