### 2025-03-08 95138b8
:white_check_mark: Add tests for generated output in docs. Track HDF5 and PNG files inside docs. Regenerate commensurability files & figure for relevant document pages. Update conditional skips and invisible code blocks in docs for appropriate testing in GH Actions.
