### 2025-03-09 b8b48ca
:construction: Tentative fix for Agama installation script.
