### 2025-03-09 f518a94
:construction: WIP: Run Agama tests on Python 3.12.
