+++ .github/workflows/full_tests.yml
@@ -121,7 +121,7 @@ jobs:
   agama:
     strategy:
       matrix:
-        python-version: [3.13]  # only test on latest Python version
+        python-version: [3.12]  # 3.13 not supported yet?
     name: Agama Scripts on Python ${{ matrix.python-version }}
     runs-on: ubuntu-latest  # no windows or macOS testing currently
     steps:
