### 2025-03-09 108ad20
:construction: WIP: Perform installation through pip (Pidgey strategy).
