### 2025-03-09 460bdb3
:construction: WIP: Setup GCC for Agama.
