### 2025-03-09 a3fa854
:wrench: Re-enable testing figures in doctests.
