+++ .github/workflows/full_tests.yml
@@ -69,11 +69,11 @@ jobs:
         shell: bash -el {0}
         run: pytest docs --ignore=docs/tessellation/implementation_ar_test.py --ignore=docs/tessellation/implementation_norm_test.py --ignore=docs/tessellation/usage/analysis/rotating_bar_test.py --ignore=docs/tessellation/usage/analysis/solar_neighborhood_test.py
 
-      # - name: Test Figures
-      #   # "bash -el {0}" required with conda-incubator/setup-miniconda@v3
-      #   # see: https://github.com/conda-incubator/setup-miniconda?tab=readme-ov-file#important
-      #   shell: bash -el {0}
-      #   run: pytest docs/tessellation/implementation_ar_test.py docs/tessellation/implementation_norm_test.py docs/tessellation/usage/analysis/rotating_bar_test.py docs/tessellation/usage/analysis/solar_neighborhood_test.py
+      - name: Test Figures
+        # "bash -el {0}" required with conda-incubator/setup-miniconda@v3
+        # see: https://github.com/conda-incubator/setup-miniconda?tab=readme-ov-file#important
+        shell: bash -el {0}
+        run: pytest docs/tessellation/implementation_ar_test.py docs/tessellation/implementation_norm_test.py docs/tessellation/usage/analysis/rotating_bar_test.py docs/tessellation/usage/analysis/solar_neighborhood_test.py
 
       # TBA: test galpy scripts inside examples
 
@@ -169,11 +169,11 @@ jobs:
         shell: bash -el {0}
         run: pytest docs --ignore=docs/tessellation/implementation_ar_test.py --ignore=docs/tessellation/implementation_norm_test.py --ignore=docs/tessellation/usage/analysis/rotating_bar_test.py --ignore=docs/tessellation/usage/analysis/solar_neighborhood_test.py
 
-      # - name: Test Figures
-      #   # "bash -el {0}" required with conda-incubator/setup-miniconda@v3
-      #   # see: https://github.com/conda-incubator/setup-miniconda?tab=readme-ov-file#important
-      #   shell: bash -el {0}
-      #   run: pytest docs/tessellation/implementation_ar_test.py docs/tessellation/implementation_norm_test.py docs/tessellation/usage/analysis/rotating_bar_test.py docs/tessellation/usage/analysis/solar_neighborhood_test.py
+      - name: Test Figures
+        # "bash -el {0}" required with conda-incubator/setup-miniconda@v3
+        # see: https://github.com/conda-incubator/setup-miniconda?tab=readme-ov-file#important
+        shell: bash -el {0}
+        run: pytest docs/tessellation/implementation_ar_test.py docs/tessellation/implementation_norm_test.py docs/tessellation/usage/analysis/rotating_bar_test.py docs/tessellation/usage/analysis/solar_neighborhood_test.py
 
       # TBA: test agama scripts inside examples
 
