### 2025-03-09 8d2038c
:construction_worker: Make figure generation conditional in full tests.
