### 2025-03-09 ac48d19
:wrench: Run examples as part of full package testing.
