+++ .github/workflows/full_tests.yml
@@ -82,7 +82,14 @@ jobs:
         shell: bash -el {0}
         run: pytest docs/tessellation/implementation_ar_test.py docs/tessellation/implementation_norm_test.py docs/tessellation/usage/analysis/rotating_bar_test.py docs/tessellation/usage/analysis/solar_neighborhood_test.py
 
-      # TBA: test galpy scripts inside examples
+      - name: Test Scripts
+        # "bash -el {0}" required with conda-incubator/setup-miniconda@v3
+        # see: https://github.com/conda-incubator/setup-miniconda?tab=readme-ov-file#important
+        shell: bash -el {0}
+        # turn on interactive mode for matplotlib, so plot windows don't halt the test
+        run: |
+          echo "interactive: true" > matplotlibrc
+          python examples/using_galpy/rotating_bar.py
 
   gala:
     strategy:
@@ -123,7 +130,14 @@ jobs:
 
       # No figures currently generated by gala
 
-      # TBA: test gala scripts inside examples
+      - name: Test Scripts
+        # "bash -el {0}" required with conda-incubator/setup-miniconda@v3
+        # see: https://github.com/conda-incubator/setup-miniconda?tab=readme-ov-file#important
+        shell: bash -el {0}
+        # turn on interactive mode for matplotlib, so plot windows don't halt the test
+        run: |
+          echo "interactive: true" > matplotlibrc
+          python examples/using_gala/rotating_bar.py
 
   agama:
     strategy:
@@ -183,7 +197,14 @@ jobs:
         shell: bash -el {0}
         run: pytest docs/tessellation/implementation_ar_test.py docs/tessellation/implementation_norm_test.py docs/tessellation/usage/analysis/rotating_bar_test.py docs/tessellation/usage/analysis/solar_neighborhood_test.py
 
-      # TBA: test agama scripts inside examples
+      - name: Test Scripts
+        # "bash -el {0}" required with conda-incubator/setup-miniconda@v3
+        # see: https://github.com/conda-incubator/setup-miniconda?tab=readme-ov-file#important
+        shell: bash -el {0}
+        # turn on interactive mode for matplotlib, so plot windows don't halt the test
+        run: |
+          echo "interactive: true" > matplotlibrc
+          python examples/using_agama/rotating_bar.py
 
   build-docs:
     name: Build Documentation with MkDocs
