### 2025-03-09 ilikecubesnstuff/commensurability#35 @ilikecubesnstuff
Currently `ro` and `zo` *are* passed in within `pidgey` using `galpy.util.conversion.get_physical`. When I try to address this warning in `pidgey`, I get this error if I change the inputs for orbit initialization:
```
AssertionError: Physical conversion for the Orbit object is not consistent with that of the Potential given to it
```

Is there a way to specify the input such that this assertion error isn't thrown AND the warning is not thrown either?
