### 2025-03-09 ilikecubesnstuff/commensurability#38 @ilikecubesnstuff
Adressed in #34.
