### 2025-03-09 ilikecubesnstuff/commensurability#35 @nstarman
Hard to address without seeing the stack trace.
SkyCoord has the galactocentric frame on it, so you can get the relevant parameters from it. And galpy's r0, v0 are arbitary, just used for unit conversions, so these should be able to be set based on the frame. But yes, it must be given consistently across all of galpy.
