### 2025-03-10 #7009 @nstarman
Thanks @ilikecubesnstuff for tackling what turned out to be a much more challenging refactor of the docs than originally anticipated. The test suite looks good and I've successfully reproduced broad portions of the docs examples (looking to the good CI setup).
Small comments remain about the Community Guidelines, but otherwise this looks good.
@dfm what is the process for wrapping up?
