### 2025-03-14 ilikecubesnstuff/commensurability#35 @ilikecubesnstuff
I don't know how to tackle this issue. The current fix in `pidgey` is to use `galpy.util.conversion.get_physical`:
```
        pot_units = self.conversion.get_physical(pot)
        orbit = self.orbit.Orbit(skycoord, **pot_units)
```
But this creates the warnings you mentioned above. I am using the script `examples/using_galpy/rotating_bar.py` to test this, and here is what I have tried:

In this case, the `skycoord` frame has the following properties: `(ra, dec) in deg (266.4051, -28.936175)>, galcen_distance=8.122 kpc, galcen_v_sun=(12.9, 245.6, 7.78) km / s, z_sun=20.8 pc, roll=0.0 deg)`

If I pass `zo` as anything other than 20.8 pc, I get the following traceback:
```
Traceback (most recent call last):
  File "/mnt/c/Users/User/Documents/GitHub/commensurability/examples/using_galpy/rotating_bar.py", line 49, in <module>
    tanal = TessellationAnalysis(
            ^^^^^^^^^^^^^^^^^^^^^
  File "/mnt/c/Users/User/Documents/GitHub/commensurability/src/commensurability/analysis.py", line 387, in __init__
    self._construct_image_with_mp(pidgey_chunksize, mp_chunksize, progressbar)
  File "/mnt/c/Users/User/Documents/GitHub/commensurability/src/commensurability/analysis.py", line 405, in _construct_image_with_mp
    orbits = self.backend.compute_orbit(
             ^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/mnt/c/Users/User/Documents/GitHub/pidgey/src/pidgey/base.py", line 94, in compute_orbit
    self._result = self._compute_orbit(
                   ^^^^^^^^^^^^^^^^^^^^
  File "/mnt/c/Users/User/Documents/GitHub/pidgey/src/pidgey/galpy_backend.py", line 33, in _compute_orbit
    orbit = self.orbit.Orbit(
            ^^^^^^^^^^^^^^^^^
  File "/home/subhadeep/miniconda3/envs/c14y_galpy/lib/python3.12/site-packages/galpy/orbit/Orbits.py", line 357, in __init__
    self._setup_parse_coordtransform(vxvv, ro, vo, zo, solarmotion, radec, lb)
  File "/home/subhadeep/miniconda3/envs/c14y_galpy/lib/python3.12/site-packages/galpy/orbit/Orbits.py", line 479, in _setup_parse_coordtransform
    raise ValueError(
ValueError: Orbit initialization's zo different from SkyCoord's z_sun; these should be the same for consistency
```

If I pass `ro` as anything other than 8.0 kpc OR `vo` as anything other than 220.0 km / s, I get the following traceback:
```
galpyWarning: Orbit's initialization normalization ro and zo are incompatible with SkyCoord's galcen_distance (should have galcen_distance^2 = ro^2 + zo^2)
Traceback (most recent call last):
  File "/mnt/c/Users/User/Documents/GitHub/commensurability/examples/using_galpy/rotating_bar.py", line 49, in <module>
    tanal = TessellationAnalysis(
            ^^^^^^^^^^^^^^^^^^^^^
  File "/mnt/c/Users/User/Documents/GitHub/commensurability/src/commensurability/analysis.py", line 387, in __init__
    self._construct_image_with_mp(pidgey_chunksize, mp_chunksize, progressbar)
  File "/mnt/c/Users/User/Documents/GitHub/commensurability/src/commensurability/analysis.py", line 405, in _construct_image_with_mp
    orbits = self.backend.compute_orbit(
             ^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/mnt/c/Users/User/Documents/GitHub/pidgey/src/pidgey/base.py", line 94, in compute_orbit
    self._result = self._compute_orbit(
                   ^^^^^^^^^^^^^^^^^^^^
  File "/mnt/c/Users/User/Documents/GitHub/pidgey/src/pidgey/galpy_backend.py", line 43, in _compute_orbit
    orbit.integrate(t, pot, **integration_kwargs)
  File "/home/subhadeep/miniconda3/envs/c14y_galpy/lib/python3.12/site-packages/galpy/orbit/Orbits.py", line 1394, in integrate
    _check_consistent_units(self, pot)
  File "/home/subhadeep/miniconda3/envs/c14y_galpy/lib/python3.12/site-packages/galpy/orbit/Orbits.py", line 8094, in _check_consistent_units
    assert physical_compatible(
           ^^^^^^^^^^^^^^^^^^^^
AssertionError: Physical conversion for the Orbit object is not consistent with that of the Potential given to it
```
