### 2025-03-15 ilikecubesnstuff/commensurability#35 @nstarman
I think you can define a galactocentric frame from the galpy and then transform the SkyCoord to that frame. Then pass it to galpy.
