### 2025-03-17 ilikecubesnstuff/commensurability#35 @ilikecubesnstuff
That does appear to produce the same warning. I am doing the following within `pidgey/galpy_backend.py`:
```py
        pot_units = self.conversion.get_physical(pot)
        skycoord_galcen = skycoord.transform_to(coord.Galactocentric())
        orbit = self.orbit.Orbit(skycoord_galcen, **pot_units)
```
Entering specific arguments into `coord.Galactocentric` appears to have no effect.
