### 2025-03-19 ilikecubesnstuff/commensurability#35 @jobovy
I'm happy to help debug this issue, I have a lot of experience with similar issues 😄 .... The issue might be that ``skycoord_galcen = skycoord.transform_to(coord.Galactocentric())`` sets the ``galcen_distance``, ``z_sun`` etc. attributes of the ``SkyCoord`` and the defaults can be incompatible with ``galpy``'s. But it would be helpful to have some additional info:
* What version of ``galpy`` are you using? (asking because I made some recent changes to how SkyCoord frame parameters are treated, see jobovy/galpy#715
* Could you create a self-contained example that leads to the issue? For example, if the code in https://github.com/ilikecubesnstuff/commensurability/issues/35#issuecomment-2730896578 is the issue, perhaps make a version of that code that starts from the potential used and the ``coord`` object?
