### 2025-03-20 ilikecubesnstuff/commensurability#35 @ilikecubesnstuff
Hi @jobovy! Thank you so much for offering to help debug this :)

I was on galpy version `1.10.1`, but I just upgraded to `1.10.2` and the warning still persists.
The follow code produces the warning in my environment with galpy `1.10.2`:
```py
import astropy.coordinates as c
import astropy.units as u
import galpy.orbit as go
import galpy.potential as gp
from galpy.util import conversion

# potential
omega = 30 * u.km / u.s / u.kpc
halo = gp.NFWPotential(conc=10, mvir=1)
disc = gp.MiyamotoNagaiPotential(amp=5e10 * u.solMass, a=3 * u.kpc, b=0.1 * u.kpc)
bar = gp.SoftenedNeedleBarPotential(
    amp=1e9 * u.solMass, a=1.5 * u.kpc, b=0 * u.kpc, c=0.5 * u.kpc, omegab=omega
)
pot = [halo, disc, bar]

# initial condition
ic = c.SkyCoord(
    x=8 * u.kpc,
    y=0 * u.kpc,
    z=1 * u.kpc,
    v_x=0 * u.km / u.s,
    v_y=220 * u.km / u.s,
    v_z=0 * u.km / u.s,
    frame="galactocentric",
    representation_type="cartesian",
)

# lines adapted from pidgey
pot_units = conversion.get_physical(pot)
ic = ic.transform_to(c.Galactocentric())  # warning appears with and without this line
orbit = go.Orbit(ic, **pot_units)
```
