### 2025-03-20 ilikecubesnstuff/commensurability#35 @jobovy
Great, thanks! Okay, so the issue is that when you initialize a ``SkyCoord`` in Galactocentric coordinates, ``astropy`` sets its default frame parameters, which are currently ``galcen_distance = 8.122 kpc`` and ``z_sun = 20.8 pc`` (and also the Solar velocity). When you set up the potential without specifying ``ro`` and ``vo``, ``galpy``'s defaults are used, which have ``ro=8 kpc`` (and ``vo=220 km/s``). So these are then incompatible (compatibility would be that ``galcen_distance^2 = ro^2 + zo^2``).

I'm not sure how ``pidgey`` does things, but here are some ways to deal with this:
1. Ignore the ``SkyCoord``'s frame parameters. I _think_ that's what is currently done and what leads to the warning, but better would be to just set the relevant ``SkyCoord`` frame parameters to the value that is compatible with ``galpy``'s. E.g., ``ic.galcen_distance = sqrt(ro**2.+zo**2.)``, where you first get ``ro`` from running ``conversion.get_physical(pot)`` as you do above. Note that you should be able to set ``ic.z_sun`` to whatever you want as long as you (i) don't explicitly set ``zo`` in the Orbit initialization and (ii) make sure ``galcen_distance`` is compatible with ``(ro,z_sun)``.
2. _First_ set up the SkyCoord, then extract its ``galcen_distance`` and set ``ro`` based on the assumed ``z_sun`` to ``ro = sqrt(galcen_distance**2.-z_sun**2.)``. And then use that ``ro`` when setting up the potential.
3. _First_ set up the potential, extract ``ro``, and use it to set up the SkyCoord. So roughly as follows:
   ```python
   rovo = conversion.get_physical(pot)
   ro=rovo["ro"]
   vo =rovo["vo"]
   # initial condition
   z_sun = 0.0208*u.kpc
   ic = c.SkyCoord(
       x=8 * u.kpc,
       y=0 * u.kpc,
       z=1 * u.kpc,
       v_x=0 * u.km / u.s,
       v_y=220 * u.km / u.s,
       v_z=0 * u.km / u.s,
       frame="galactocentric",
       representation_type="cartesian",
       galcen_distance=numpy.sqrt(ro**2.*u.kpc**2 + z_sun**2),
       z_sun=z_sun
   )
   ```
   That gets rid of the warning.

Note that there isn't an issue with the assumed solar velocity, because ``astropy`` sets the solar velocity directly, while ``galpy`` splits it into circular velocity + ``solarmotion`` and if you don't explicitly set ``solarmotion``, then they are just made consistent for any ``vo``.

Ideally, at some point ``galpy`` would not use ``ro,vo`` for both unit conversion and the Solar <-> Galactocentric frame conversion, but there is no plan to change this currently.
