### 2025-03-21 ilikecubesnstuff/commensurability#35 @ilikecubesnstuff
Thank you so much! For `pidgey`, I think the best option is defining a new `SkyCoord` object based on the one passed in, just with the extra keyword arguments for consistency in `galpy`. (`commensurability` defines the potential and initial conditions separately and then passes both in.)
