### 2025-03-21 ilikecubesnstuff/commensurability#35 @ilikecubesnstuff
Changes addressing this issue are now merged in `pidgey` and `commensurability`.
