### 2025-04-12 #7009 @dfm
@ilikecubesnstuff — I've opened a small PR with some tiny edits to the bibliography, please take a look and merge or let me know what you think.

Once you've done that:
