+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "commensurability"
-version = "1.0.0"
+version = "1.1.0"
 description = "A package for analyzing orbit commensurabilities."
 authors = [
     { name = "ilikecubesnstuff", email = "25328250+ilikecubesnstuff@users.noreply.github.com" },
