### 2025-05-06 #7009 @ilikecubesnstuff
Hi @dfm! I am running into some issues with installing Docker and generating the PDF. I asked my co-author to help with this, since he generated the PDF last time. Ideally, I should be able to report back a DOI this week.
