+++ paper/paper.md
@@ -14,12 +14,12 @@ authors:
     affiliation: "1, 2"
     orcid: 0009-0006-9454-5141
   - name: Michael S. Petersen
-    affiliation: "1"
+    affiliation: "2"
     orcid: 0000-0003-1517-3935
 affiliations:
- - name: Insitute for Astronomy, University of Edinburgh
-   index: 1
  - name: Institut de Ciències del Cosmos, Universitat de Barcelona
+   index: 1
+ - name: Insitute for Astronomy, University of Edinburgh
    index: 2
 date:
 bibliography: paper.bib
