### 2025-05-10 #7009 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
