### 2025-05-10 e924f2f
Initial editorial tweaks for JOSS paper
