+++ paper/paper.md
@@ -5,7 +5,7 @@ tags:
   - astronomy
   - dynamics
   - galactic dynamics
-  - milky way
+  - Milky Way
   - tessellation
 
 authors:
@@ -17,9 +17,9 @@ authors:
     affiliation: "2"
     orcid: 0000-0003-1517-3935
 affiliations:
- - name: Institut de Ciències del Cosmos, Universitat de Barcelona
+ - name: Institut de Ciències del Cosmos, Universitat de Barcelona, Spain
    index: 1
- - name: Insitute for Astronomy, University of Edinburgh
+ - name: Institute for Astronomy, University of Edinburgh, UK
    index: 2
 date:
 bibliography: paper.bib
@@ -30,7 +30,7 @@ arxiv-doi:
 
 # Summary
 
-Stars like our Sun orbit the center of our Milky Way galaxy. Over many orbits in the stellar disk of the galaxy, a star will (hypothetically) probe every point in a 3D toroid defined by a minimum/maximum radius and minimum/maximum height above the midplane. However, there are special classes of orbits that repeat their tracks over many revolutions around the Galactic center, and therefore only probe a small sub-volume of the toroid. This property stems from low-integer orbital frequency ratios -- commensurate frequencies -- and such orbits are referred to as satisfying a "commensurability".
+Stars like our Sun orbit the center of our Milky Way galaxy. Over many orbits in the stellar disk of the galaxy, a star will (hypothetically) probe every point in a 3D toroid defined by a minimum/maximum radius and minimum/maximum height above the midplane. However, there are special classes of orbits that repeat their tracks over many revolutions around the Galactic center, and therefore only probe a small sub-volume of the toroid. This property stems from low-integer orbital frequency ratios—commensurate frequencies—and such orbits are referred to as satisfying a "commensurability".
 
 To study the orbital content of galaxies, astronomers often use models for galaxies that allow for the integration of orbits within the model galaxy's potential. The integration results in a time series of positions (e.g. $x,y,z$) for the orbit that can then be analyzed to produce a classification. Complicating this picture, potentials evolving with time introduce a new frequency: the pattern frequency of the evolution. The evolution of the potential might be sourced by the rotation of a galactic bar, the formation or dissolution of spiral arms, the growth of the galaxy, or the interaction with a satellite galaxy. Orbits that are commensurate with the pattern frequency are particularly important in galactic dynamics as they experience the same force fluctuations during every revolution: this causes changes to the typically conserved quantities of orbits (e.g. energy and angular momentum).
 
@@ -43,7 +43,7 @@ The Python package presented here, `commensurability`, provides a straightforwar
 
 The technique of measuring orbit volumes has previously been used to classify orbits in 2D in both fixed and evolving potentials, revealing families of commensurate orbits that are critical for galactic evolution [@Petersen.Weinberg.Katz.2021]. However, the technique had not (1) been extended to 3D, and (2) did not have a user-friendly Python interface. The `commensurability` package solves both problems and provides interoperability with three leading galactic dynamics packages to accomplish the underlying orbit integrations.
 
[…]
