### 2025-05-10 #7009 @warrickball
Hi @ilikecubesnstuff! I've opened [a small pull request](https://github.com/ilikecubesnstuff/commensurability/pull/46) with some editorial tweaks (and a suggestion to remove a section). If you're happy with those, merge them and make any other final changes, I'll launch another test run of the publication pipeline to see if we're ready to publish.
