### 2025-05-13 #7009 @ilikecubesnstuff
Hi @warrickball, many apologies for opening this up again—I was informed by my supervisors that I must include a line about my funding in the acknowledgements section of the paper. I did add it to the Zenodo archive, but I did not make the corresponding edit on the JOSS paper itself. Is it possible to make an amendment?
