### 2025-05-13 #7009 @ilikecubesnstuff
@warrickball thank you so much—I've remade a `joss` branch and added the acknowledgement there: https://github.com/ilikecubesnstuff/commensurability/commit/33258b78ea62c8c3e85dd38578319db808c0bdd5
