### 2025-05-13 #7009 @ilikecubesnstuff
I've updated the PDF in my repository—thank you so much once again @warrickball. 😄
Do you know when the document in https://joss.theoj.org/papers/10.21105/joss.07009 will get updated?
