### 2025-05-13 #7009 @warrickball
> I've updated the PDF in my repository—thank you so much once again [@warrickball](https://github.com/warrickball). 😄

No problem!

> Do you know when the document in https://joss.theoj.org/papers/10.21105/joss.07009 will get updated?

I'm not sure why it took longer than the initial acceptance step but the PDF at the DOI looks up to date to me now.
